Add files via upload
1) Глобальные переменные разместил в классах, которые их используют
2) Вытащил из "grafic_editor" все лишние классы, теперь там только подключаются непосредственно режимы отрисовки. Проверил, что все режимы работают автономно
3) Переименовал: lcm -> mouse_left, rcm -> mouse_right.
