+++ eraser.cpp
@@ -1,0 +1,6 @@
+#pragma once
+#include"eraser.h"
+
+eraser::eraser() {
+	size = 4;
+}
